pyproject.toml
[project]
name = "multi-sdk-llm-notebooks"
version = "0.1.0"
description = "Side-by-side OpenAI SDK vs Anthropic SDK benchmarks for function calling and structured output."
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
  "anthropic>=0.40.0",
  "openai>=1.50.0",
  "python-dotenv>=1.0.1",
  "pandas>=2.2.0",
  "tabulate>=0.9.0",
  "pydantic>=2.9.0",
  "openai-agents>=0.14.6",
  "nest-asyncio>=1.6.0",
]

[dependency-groups]
dev = [
  "jupyterlab>=4.2.0",
  "jupytext>=1.16.4",
  "ipykernel>=6.29.5",
  "matplotlib>=3.9.0",
]

[tool.jupytext]
formats = "ipynb,py:percent"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
